Fixed Segmentation Fault, Fixed nextDisk
A segmentation fault occurred due to missing tracks for the seeding region. Now skipping events with missing tracks.

The for loop in the nextDisk() function did not account for the case where, in the forward propagation, the track moves from a scintillator tile to a silicon cell. In the context of the for loop this requires 3 iterations instead of two to move to the silicon in the next layer (scintillator, silicon, scintillator, silicon).